Allow no outdir if dryrunning and ensure log file location
Stop getpapers from exiting if no output directory is given
and we are dry running. This means logging must happen outside of
the output directory. Logging was only happening in the output
directory because the logger didn't usually initialise as before
we changed directory. Now we ensure we know where we are writing by
getting a writable file stream and passing that straight to winston